% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{albums}
\alias{albums}
\title{An artist's music albums.}
\usage{
albums(name, message = TRUE)
}
\arguments{
\item{name}{The name of the artist/band.}

\item{message}{Should the function print something if the
required data is not found?}
}
\value{
\code{gente} returns a data.frame with information
    about the albums, as the id, name and year of release.
}
\description{
Gives information about the albums  of an artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

\dontrun{
albums("the-beatles")
albums("chico-buarque")
}
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
