% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artistInfo.R
\name{artistInfo}
\alias{artistInfo}
\title{Artist Information}
\usage{
artistInfo(name, message = TRUE)
}
\arguments{
\item{name}{The name of the artist/band.}

\item{message}{Should the function print something if the
required data is not found?}
}
\value{
\code{artistInfo} returns a data.frame with the information.
}
\description{
Gives some information about a given artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

\dontrun{
artistInfo("the-beatles")
artistInfo("chico-buarque")
}

}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
