% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography}
\alias{discography}
\title{Discography of an artist/band.}
\arguments{
\item{name}{The name of the artist/band.}
}
\value{
\code{discography} returns a data.frame with information
    about the artist and the discography.
}
\description{
Gives the list of albums of an specific artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

discography("the-beatles")

}
\author{
Bruna Wundervald, \email{brunadaviesw@ufpr.br}.
}
