% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre.R
\name{genre}
\alias{genre}
\title{An artist's musical genre(s)}
\usage{
genre(name, message = TRUE)
}
\arguments{
\item{name}{The name of the artist/band.}

\item{message}{Should the function print something if the
required data is not found?}
}
\value{
\code{genre} returns a data.frame with information
    about the genre(s).
}
\description{
Gives information about the genre (ou multiple genres)
of an artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

\dontrun{
genre("the-beatles")
genre("chico-buarque")
}
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
