% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedInfo.R
\name{relatedInfo}
\alias{relatedInfo}
\title{Artist's Related}
\arguments{
\item{name}{The name of the artist/band.}
}
\value{
\code{relatedInfo} returns a data.frame with information
    about the related artists.
}
\description{
Gives information about what artists/bands are
    related to a specific artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

relatedInfo("the-beatles")
relatedInfo("chico-buarque")

}
\author{
Bruna Wundervald, \email{brunadaviesw@ufpr.br}.
}
