% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songNames.R
\name{songNames}
\alias{songNames}
\title{Song names of an artist/band.}
\usage{
songNames(name)
}
\arguments{
\item{name}{The name of the artist/band.}
}
\value{
\code{relatedInfo} returns a data.frame with information
    about song names.
}
\description{
Gives information about the song names of an specific
    artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

\dontrun{
songNames("the-beatles")
songNames("chico-buarque")
}

}
\author{
Bruna Wundervald, \email{brunadaviesw@ufpr.br}.
}
