% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topLyrics.R
\name{topLyrics}
\alias{topLyrics}
\title{Top lyrics of an artist/band}
\usage{
topLyrics(name)
}
\arguments{
\item{name}{The name of the artist/band.}
}
\value{
\code{topLyrics} returns a data.frame with information
    about the top lyrics.
}
\description{
Gives information about the top lyrics (most viewed)
    about an specific artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

\dontrun{
topLyrics("the-beatles")
topLyrics("chico-buarque")
}

}
\author{
Bruna Wundervald, \email{brunadaviesw@ufpr.br}.
}
