% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography}
\alias{discography}
\title{Discography of an artist/band.}
\usage{
discography(name, message = TRUE)
}
\arguments{
\item{name}{The name of the artist/band.}

\item{message}{Should the function print something if the
required data is not found?}
}
\value{
\code{discography} returns a data.frame with information
    about the artist and the discography.
}
\description{
Gives the list of albums of an specific artist/band.
}
\details{
The variables returned by the function are extracted with
    the Vagalume API.
}
\examples{

\dontrun{
discography("the-beatles")
}
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
